env count changed to size_t; env table key changed to uint64_t; add clang-tidy configs

+++ src/include/env.h
@@ -1,6 +1,7 @@
 #ifndef ENV_H
 #define ENV_H
 
+#include <stddef.h>
 #ifdef __cplusplus
 extern "C" {
 #endif
@@ -11,13 +12,12 @@ extern "C" {
 // Environment Structure
 typedef struct PEnv {
     // Variable Table
-    // We are using Array for `now`. Note: Switch to Hash table
     struct {
-        uint32_t key;
+        uint64_t key;
         PValue value;
     } *table;
     // Count of pairs in table
-    int count;
+    size_t count;
     // Parent Environment
     struct PEnv *enclosing;
 } PEnv;
@@ -37,16 +37,16 @@ void FreeEnv(PEnv *e);
 void DebugEnv(PEnv *e);
 
 // Create a Pair and push to the Environment table
-void EnvPutValue(PEnv *e, uint32_t hash, PValue value);
+void EnvPutValue(PEnv *e, uint64_t hash, PValue value);
 
 // Fetch value from table corresponding the provided hash
 // Returns `NULL` if not found.
-PValue EnvGetValue(PEnv *e, uint32_t hash, bool *found);
+PValue EnvGetValue(PEnv *e, uint64_t hash, bool *found);
 
 // Set value to existing variable in table;
 // Returns true if exists and was successful in updating value;
 // Returns false if key doesn't exist, or failed to update.
-bool EnvSetValue(PEnv *e, uint32_t hash, PValue value);
+bool EnvSetValue(PEnv *e, uint64_t hash, PValue value);
 
 #ifdef __cplusplus
 }
